Add optional requestHeaders argument to autoUpdater.setFeedURL + allow calling mulitple times
