+++ atom/browser/auto_updater.h
@@ -5,6 +5,7 @@
 #ifndef ATOM_BROWSER_AUTO_UPDATER_H_
 #define ATOM_BROWSER_AUTO_UPDATER_H_
 
+#include <map>
 #include <string>
 
 #include "base/macros.h"
@@ -42,11 +43,14 @@ class Delegate {
 
 class AutoUpdater {
  public:
+  typedef std::map<std::string, std::string> HeaderMap;
+
   // Gets/Sets the delegate.
   static Delegate* GetDelegate();
   static void SetDelegate(Delegate* delegate);
 
-  static void SetFeedURL(const std::string& url);
+  static void SetFeedURL(const std::string& url,
+                         const HeaderMap& requestHeaders);
   static void CheckForUpdates();
   static void QuitAndInstall();
