+++ lib/browser/api/auto-updater/auto-updater-native.js
@@ -3,4 +3,8 @@ const autoUpdater = process.atomBinding('auto_updater').autoUpdater
 
 Object.setPrototypeOf(autoUpdater, EventEmitter.prototype)
 
+autoUpdater.setFeedURL = function (url, headers) {
+  return autoUpdater._setFeedURL(url, headers || {})
+}
+
 module.exports = autoUpdater
