+++ lib/browser/api/auto-updater/auto-updater-win.js
@@ -19,7 +19,7 @@ AutoUpdater.prototype.quitAndInstall = function () {
   return app.quit()
 }
 
-AutoUpdater.prototype.setFeedURL = function (updateURL) {
+AutoUpdater.prototype.setFeedURL = function (updateURL, headers) {
   this.updateURL = updateURL
 }
